requirements.txt
gspread
oauth2client
pandas
streamlit
google-api-python-client
google-auth
requests
python-dotenv
pyngrok
beautifulsoup4
jinja2